requirements.txt
faster-whisper>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
silero-vad>=5.0.0
structlog>=24.0.0
typer>=0.9.0
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
streamlit>=1.28.0
pyinstaller>=6.0.0
keyring>=24.0.0
pystray>=0.19.0
pillow>=10.0.0
psutil>=5.9.0

nvidia-cudnn-cu12>=9.0.0
nvidia-cublas-cu12>=12.0.0
